# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin": {
      "get": {
        "operationId": "AdminController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Admin"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/{id}": {
      "get": {
        "operationId": "AdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admin"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "AdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admin"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/class": {
      "post": {
        "operationId": "ClassController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Class"
                }
              }
            }
          }
        },
        "tags": [
          "Class"
        ]
      },
      "get": {
        "operationId": "ClassController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Class"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Class"
        ]
      }
    },
    "/api/class/{id}": {
      "get": {
        "operationId": "ClassController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Class"
                }
              }
            }
          }
        },
        "tags": [
          "Class"
        ]
      },
      "put": {
        "operationId": "ClassController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Class"
                }
              }
            }
          }
        },
        "tags": [
          "Class"
        ]
      },
      "delete": {
        "operationId": "ClassController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Class"
        ]
      }
    },
    "/api/parent-or-guardians": {
      "post": {
        "operationId": "ParentOrGuardiansController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParentOrGuardianDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentOrGuardian"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Parent & Guardians"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ParentOrGuardiansController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ParentOrGuardian"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Parent & Guardians"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/parent-or-guardians/{id}": {
      "get": {
        "operationId": "ParentOrGuardiansController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentOrGuardian"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Parent & Guardians"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "ParentOrGuardiansController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateParentOrGuardianDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentOrGuardian"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Parent & Guardians"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ParentOrGuardiansController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Parent & Guardians"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/academic-year": {
      "post": {
        "operationId": "AcademicYearController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAcademicYearDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYear"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Academic Year"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "AcademicYearController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AcademicYear"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Academic Year"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/academic-year/{id}": {
      "get": {
        "operationId": "AcademicYearController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYear"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Academic Year"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "AcademicYearController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAcademicYearDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYear"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Academic Year"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AcademicYearController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Academic Year"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "School Compass 365",
    "description": "swagger json: <a href=\"/swagger.json\">swagger.json</a>",
    "version": "1.0",
    "contact": {},
    "license": {
      "name": "Mozilla Public License Version 2.0",
      "url": "https://github.com/vaibhavdubay/School-Compass-365-be/blob/main/LICENSE"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "enum": [
              "super_admin",
              "admin",
              "teacher",
              "student"
            ],
            "type": "string"
          },
          "changePassword": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "userName",
          "password",
          "role",
          "changePassword",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "AcademicYear": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "academicYear": {
            "type": "string"
          },
          "current": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "academicYear",
          "current",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Image": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "filename": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "super_admin",
              "admin",
              "teacher",
              "student"
            ]
          },
          "mimeType": {
            "type": "string"
          },
          "buffer": {
            "type": "object"
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "role",
          "mimeType",
          "buffer",
          "userId",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Class": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "className": {
            "type": "string"
          },
          "nextClass": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "streamsRequired": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "className",
          "nextClass",
          "order",
          "streamsRequired",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "School": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "establishedYear": {
            "type": "number"
          },
          "address1": {
            "type": "string"
          },
          "address2": {
            "type": "string"
          },
          "academicYears": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AcademicYear"
            }
          },
          "logo": {
            "$ref": "#/components/schemas/Image"
          },
          "logoUrl": {
            "type": "string"
          },
          "currentAcademicYear": {
            "$ref": "#/components/schemas/AcademicYear"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "pincode": {
            "type": "number"
          },
          "schoolDISECode": {
            "type": "string"
          },
          "schoolCode": {
            "type": "string"
          },
          "classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Class"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phoneNumber",
          "establishedYear",
          "address1",
          "address2",
          "academicYears",
          "logo",
          "logoUrl",
          "currentAcademicYear",
          "city",
          "state",
          "pincode",
          "schoolDISECode",
          "schoolCode",
          "classes",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Admin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "school": {
            "$ref": "#/components/schemas/School"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "gender": {
            "enum": [
              "male",
              "female",
              "other"
            ],
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "profileImage": {
            "$ref": "#/components/schemas/Image"
          },
          "profileImageUrl": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "school",
          "firstName",
          "lastName",
          "gender",
          "email",
          "phoneNumber",
          "profileImage",
          "profileImageUrl",
          "user",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "UpdateAdminDto": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "format": "binary"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "gender",
          "email",
          "phoneNumber"
        ]
      },
      "ParentOrGuardian": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "gender": {
            "enum": [
              "male",
              "female",
              "other"
            ],
            "type": "string"
          },
          "relations": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "gender",
          "relations",
          "contactEmail",
          "contactPhone",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "CreateClassDto": {
        "type": "object",
        "properties": {
          "className": {
            "type": "string"
          }
        },
        "required": [
          "className"
        ]
      },
      "UpdateClassDto": {
        "type": "object",
        "properties": {
          "className": {
            "type": "string"
          }
        }
      },
      "CreateParentOrGuardianDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "relations": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "gender",
          "relations",
          "contactEmail",
          "contactPhone"
        ]
      },
      "UpdateParentOrGuardianDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "relations": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          }
        }
      },
      "CreateAcademicYearDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateAcademicYearDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}